# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PLoT Engine API
  version: 1.0.0
  description: |
    PLoT Engine v1 - Deterministic, auditable probabilistic models with trust signals.
    
    **Trust Signals**: All /v1 responses include Model Card, Confidence badge, and Explain-Δ.
    
    **Demo Mode**: Add `?demo=1` or header `X-Demo: 1` for deterministic sample payloads.
    When present, demo mode short-circuits request validation and returns a canonical demo payload.
    
    **Feature Flags**:
    - FEATURE_STREAM=1 enables real SSE /stream route (test-only otherwise)
    - AUTH_ENABLED=1 requires Authorization: Bearer <token> on certain routes
    
    **Legacy Routes**: /draft-flows, /critique, /improve (v0) remain for backwards compatibility.
  contact:
    name: PLoT Engine
    url: https://github.com/Talchain/plot-lite-service
    email: devnull@talchain.ai
servers:
  - url: http://127.0.0.1:4311
    description: Local development
  - url: https://api.olumi.ai
    description: Production (future)
...
paths:
...
  /v1/evidence:
    post:
      summary: Evidence & Priors
      description: |
        Apply reference-class priors to influence beliefs.
        Returns adjusted inference with baseline comparison.
        Note: v1 uses simple weighted average; full Bayesian update in future versions.
      operationId: evidence
      tags: [v1, inference]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [graph, priors]
              properties:
                graph:
                  type: object
                  required: [nodes, edges]
                  properties:
                    nodes:
                      type: array
                      items:
                        type: object
                        required: [id]
                        properties:
                          id:
                            type: string
                          label:
                            type: string
                    edges:
                      type: array
                seed:
                  type: integer
                  example: 4242
                priors:
                  type: array
                  items:
                    type: object
                    required: [node_id, mean, variance]
                    properties:
                      node_id:
                        type: string
                      mean:
                        type: number
                      variance:
                        type: number
                      source:
                        type: string
                        example: "refclass:xyz"
            example:
              graph:
                nodes:
                  - {id: "A", label: "Node A"}
                edges: []
              seed: 4242
              priors:
                - {node_id: "A", mean: 0.8, variance: 0.05, source: "refclass:xyz"}
      responses:
        '200':
          description: Evidence results
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: string
                    example: "evidence.v1"
                  result:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          p10:
                            type: number
                          p50:
                            type: number
                          p90:
                            type: number
                      baseline:
                        type: object
                        properties:
                          p50:
                            type: number
                      adjustment:
                        type: object
                        properties:
                          delta:
                            type: number
                          direction:
                            type: string
                            enum: [increase, decrease, none]
                  top_drivers:
                    type: array
                  meta:
                    type: object
                    properties:
                      seed:
                        type: integer
                      inference_mode:
                        type: string
                      prior_alpha:
                        type: number
                      limitations:
                        type: string
              example:
                schema: "evidence.v1"
                result:
                  summary: {p10: 0.524, p50: 0.724, p90: 0.924}
                  baseline: {p50: 0.624}
                  adjustment: {delta: 0.1, direction: "increase"}
                top_drivers:
                  - {node_id: "A", contribution: 80, sign: "+", source: "refclass:xyz"}
                meta: 
                  seed: 4242
                  inference_mode: "model_based"
                  prior_alpha: 0.3
                  limitations: "v1: Simple weighted average. Full Bayesian update in future versions."
        '400':
          description: Bad input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                      message:
                        type: string
              examples:
                invalid_prior:
                  value:
                    schema: "error.v1"
                    code: "BAD_INPUT"
                    message: "Invalid node_ids in priors: Z"
...
  /v1/optimise:
    post:
      summary: Action Optimiser
      description: Select actions under budget to maximize utility with optional constraints
      operationId: optimise
      tags: [v1, optimisation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [graph, budget, actions, objective]
              properties:
                graph: { $ref: '#/components/schemas/graph' }
                budget: { type: number, minimum: 0 }
                actions:
                  type: array
                  items:
                    type: object
                    required: [id, cost, do]
                    properties:
                      id: { type: string }
                      cost: { type: number, minimum: 0 }
                      do:
                        type: array
                        items:
                          type: object
                          properties:
                            node_id: { type: string }
                            set_to: { type: number }
                objective:
                  type: object
                  properties:
                    type: { type: string, enum: [utility_linear] }
                    weights: { type: object }
                seed: { type: integer }
                constraints:
                  type: object
                  description: Optional constraints on action selection
                  properties:
                    must:
                      type: array
                      items: { type: string }
                      description: Action IDs that must be selected
                    must_not:
                      type: array
                      items: { type: string }
                      description: Action IDs that must not be selected
                    max_changed_nodes:
                      type: integer
                      minimum: 0
                      description: Maximum number of nodes that can be affected
            example:
              graph:
                nodes:
                  - id: "revenue"
                    label: "Revenue"
                  - id: "cost"
                    label: "Cost"
                edges: []
              budget: 100
              actions:
                - id: "increase_marketing"
                  cost: 50
                  do:
                    - node_id: "revenue"
                      set_to: 1.2
                - id: "reduce_overhead"
                  cost: 30
                  do:
                    - node_id: "cost"
                      set_to: 0.8
              objective:
                type: "utility_linear"
                weights:
                  revenue: 1.0
                  cost: -0.5
              seed: 4242
      responses:
        '200':
          description: Optimisation results
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema: { type: string, example: "optimise.v1" }
                  selected: { type: array, items: { type: string } }
                  utility: { type: object }
                  explanations: { type: array }
                  meta: { type: object }
        '400':
          description: Bad input or infeasible constraints
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      schema: { type: string, example: "error.v1" }
                      code: { type: string, enum: [BAD_INPUT, INFEASIBLE] }
                      message: { type: string }
                      violations:
                        type: array
                        items:
                          type: object
                          properties:
                            type: { type: string }
                            message: { type: string }
                            constraint: { type: string }
                            node_id: { type: string }
                            required: { type: number }
                            available: { type: number }
                            reason: { type: string }
              examples:
                infeasible:
                  value:
                    schema: "error.v1"
                    code: "INFEASIBLE"
                    message: "Constraints cannot be satisfied"
...
components:
  schemas:
...
    graph:
      type: object
      required: [nodes, edges]
      properties:
        version:
          type: string
          description: 'Template version (e.g., "1.2")'
        default_seed:
          type: integer
          minimum: 1
          description: 'Default seed for reproducibility'
        nodes:
          type: array
          maxItems: 50
          minItems: 1
          items:
            type: object
            required: [id, label]
            properties:
              id: { type: string, maxLength: 100 }
              label: { type: string, maxLength: 200 }
              body: { type: string, maxLength: 200, description: 'Decision context or description' }
              type: { type: string, maxLength: 50, deprecated: true }
              kind:
                type: string
                enum: [goal, decision, option, outcome]
                description: 'Node type in decision graph'
              prior:
                type: number
                minimum: 0
                maximum: 1
                description: 'Prior probability for option nodes [0..1]'
              utility:
                type: number
                minimum: -1
                maximum: 1
                description: 'Utility/payoff for outcome nodes [-1..+1]'
              effect:
                type: object
                description: 'Optional non-linear effect transform (backwards-compatible)'
                properties:
                  kind:
                    type: string
                    enum: [linear, threshold, piecewise_linear]
                    description: 'Effect type (default: linear)'
                  params:
                    type: object
                    description: 'Effect-specific parameters'
                examples:
                  - kind: linear
                  - kind: threshold
                    params: {threshold: 0.6, low: 0.0, high: 1.0}
                  - kind: piecewise_linear
                    params:
                      breakpoints:
                        - {x: 0.0, y: 0.0}
                        - {x: 0.5, y: 0.8}
                        - {x: 1.0, y: 1.0}
        edges:
          type: array
          maxItems: 200
          items:
            type: object
            required: [from, to]
            properties:
              from: { type: string, maxLength: 100 }
              to: { type: string, maxLength: 100 }
              label: { type: string, maxLength: 200 }
              weight: { type: number, minimum: -1000000, maximum: 1000000 }
              belief:
                type: number
                minimum: 0
                maximum: 1
                description: 'Inclusion confidence [0..1]. Ingress accepts legacy confidence/probability (mapped to belief). Templates emit default belief=1.0.'
              provenance:
                type: string
                maxLength: 100
                description: 'Source attribution (e.g., "template", "assumption", "user")'
              confidence:
                type: number
                minimum: 0
                maximum: 1
                deprecated: true
                description: 'DEPRECATED: Use belief instead. Accepted on ingress, never emitted.'
              probability:
                type: number
                minimum: 0
                maximum: 1
                deprecated: true
                description: 'DEPRECATED: Use belief instead. Accepted on ingress, never emitted.'
        meta:
          type: object
          description: 'Template metadata'
          properties:
            roots:
              type: array
              items: { type: string }
              description: 'Root node IDs'
            leaves:
              type: array
              items: { type: string }
              description: 'Leaf node IDs'
            suggested_positions:
              type: object
              additionalProperties:
                type: object
                required: [x, y]
                properties:
                  x: { type: number }
                  y: { type: number }
...